move/#94 : 비밀번호 변경 기능 Login -> MyInfo 모듈로 이동
