+++ feature/login/src/main/java/com/suwiki/feature/login/navigation/LoginNavigation.kt
@@ -6,7 +6,6 @@ import androidx.navigation.compose.composable
 import com.suwiki.feature.login.findid.FindIdRoute
 import com.suwiki.feature.login.findpassword.FindPasswordRoute
 import com.suwiki.feature.login.login.LoginRoute
-import com.suwiki.feature.login.resetpassword.ResetPasswordRoute
 
 fun NavController.navigateLogin() {
   navigate(LoginRoute.route)
@@ -20,16 +19,11 @@ fun NavController.navigateFindPassword() {
   navigate(LoginRoute.findPasswordRoute)
 }
 
-fun NavController.navigateResetPassword() {
-  navigate(LoginRoute.resetPasswordRoute)
-}
-
 fun NavGraphBuilder.loginNavGraph(
   popBackStack: () -> Unit,
   navigateFindId: () -> Unit,
   navigateFindPassword: () -> Unit,
   navigateSignup: () -> Unit,
-  navigateLogin: () -> Unit,
   handleException: (Throwable) -> Unit,
 ) {
   composable(route = LoginRoute.route) {
@@ -55,20 +49,10 @@ fun NavGraphBuilder.loginNavGraph(
       handleException = handleException,
     )
   }
-
-  composable(route = LoginRoute.resetPasswordRoute) {
-    ResetPasswordRoute(
-      popBackStack = popBackStack,
-      navigateFindPassword = navigateFindPassword,
-      navigateLogin = navigateLogin,
-      handleException = handleException,
-    )
-  }
 }
 
 object LoginRoute {
   const val route = "login"
   const val findIdRoute = "find-id"
   const val findPasswordRoute = "find-password"
-  const val resetPasswordRoute = "reset-password"
 }
