+++ feature/navigator/src/main/java/com/suwiki/feature/navigator/MainNavigator.kt
@@ -16,9 +16,9 @@ import com.suwiki.feature.lectureevaluation.viewerreporter.navigation.navigateLe
 import com.suwiki.feature.login.navigation.navigateFindId
 import com.suwiki.feature.login.navigation.navigateFindPassword
 import com.suwiki.feature.login.navigation.navigateLogin
-import com.suwiki.feature.login.navigation.navigateResetPassword
 import com.suwiki.feature.myinfo.navigation.navigateMyAccount
 import com.suwiki.feature.myinfo.navigation.navigateMyInfo
+import com.suwiki.feature.myinfo.navigation.navigateResetPassword
 import com.suwiki.feature.notice.navigation.navigateNotice
 import com.suwiki.feature.notice.navigation.navigateNoticeDetail
 import com.suwiki.feature.openmajor.navigation.navigateOpenMajor
