+++ feature/navigator/src/main/java/com/suwiki/feature/navigator/MainScreen.kt
@@ -68,7 +68,6 @@ internal fun MainScreen(
           navigateFindId = navigator::navigateFindId,
           navigateFindPassword = navigator::navigateFindPassword,
           navigateSignup = navigator::navigateSignup,
-          navigateLogin = navigator::navigateLogin,
           handleException = viewModel::handleException,
         )
 
@@ -135,6 +134,8 @@ internal fun MainScreen(
           navigateMyAccount = navigator::navigateMyAccount,
           navigateResetPassword = navigator::navigateResetPassword,
           navigateQuit = { /*TODO*/ },
+          navigateFindPassword = navigator::navigateFindPassword,
+          navigateLogin = navigator::navigateLogin,
           handleException = viewModel::handleException,
         )
